requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.1.31
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
magic-filter==1.0.12
matplotlib==3.10.1
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pillow==11.3.0
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
six==1.17.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
pytest~=8.3.5
pytest-asyncio
aiosqlite~=0.21.0
plotly~=6.1.2
kaleido
Telethon~=1.40.0
Faker~=37.4.0
allure-pytest